backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pdfplumber==0.11.0
openpyxl==3.1.2
pillow>=10.0.0
pandas>=2.0.0
python-dateutil>=2.8.0
requests>=2.31.0
google-api-python-client>=2.110.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
